refactor: move result formatting logic to separate utility file and update imports
